pyproject.toml
[tool.poetry]
authors = ["veeso <christian.visintin1997@gmail.com>"]
description = "A python CLI tool to extract voice sentences from audio files with speech recognition"
documentation = "https://veeso.github.io/vxt"
homepage = "https://veeso.github.io/vxt"
license = "MIT"
maintainers = ["veeso <christian.visintin1997@gmail.com>"]
name = "vxt"
readme = "README.md"
repository = "https://github.com/veeso/vxt"
version = "0.1.0"
keywords = [
    "Speech recognition",
    "voice extract",
    "audio voice",
    "split audio"
]

[tool.poetry.dependencies]
python = "^3.7"
pydub = "^0.25.0"
yaspin = "^2.1.0"
inquirer = "^2.8.0"
SpeechRecognition  = "^3.8.0"
termcolor = "^1.1.0"

[tool.poetry.dev-dependencies]
pytest = "^6.2.5"
flake8 = "^4.0.1"
flake8-bugbear = "^21.11.29"
black = "^21.12b0"

[tool.poetry.urls]
issues = "https://github.com/veeso/vxt/issues"

# Coverage setup

[tool.coverage.run]
omit = [".*", "*/site-packages/*"]

[tool.coverage.report]
fail_under = 10 # Will fail under this threshold

# Build

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
